consolidated data modules for heap'
